MLH-632 use limit in edge fetching itself

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v1/DeleteHandlerV1.java
@@ -46,6 +46,7 @@ import org.apache.atlas.repository.graphdb.janus.AtlasJanusGraph;
 import org.apache.atlas.repository.store.graph.v2.AtlasGraphUtilsV2;
 import org.apache.atlas.repository.store.graph.v2.AtlasRelationshipStoreV2;
 import org.apache.atlas.repository.store.graph.v2.EntityGraphRetriever;
+import org.apache.atlas.repository.store.graph.v2.tags.TagDAO;
 import org.apache.atlas.repository.store.graph.v2.tasks.ClassificationTask;
 import org.apache.atlas.repository.store.graph.v2.tasks.TaskUtil;
 import org.apache.atlas.tasks.TaskManagement;
@@ -80,7 +81,7 @@ import static org.apache.atlas.repository.store.graph.v2.AtlasGraphUtilsV2.*;
 import static org.apache.atlas.repository.store.graph.v2.tasks.ClassificationPropagateTaskFactory.CLASSIFICATION_PROPAGATION_ADD;
 import static org.apache.atlas.repository.store.graph.v2.tasks.ClassificationPropagateTaskFactory.CLASSIFICATION_PROPAGATION_DELETE;
 import static org.apache.atlas.repository.store.graph.v2.tasks.ClassificationPropagateTaskFactory.CLASSIFICATION_REFRESH_PROPAGATION;
-import static org.apache.atlas.repository.store.graph.v2.tasks.ClassificationTask.PARAM_ENTITY_GUID;
+import static org.apache.atlas.repository.store.graph.v2.tasks.ClassificationTask.*;
 import static org.apache.atlas.type.AtlasStructType.AtlasAttribute.AtlasRelationshipEdgeDirection.OUT;
 import static org.apache.atlas.type.Constants.HAS_LINEAGE;
 import static org.apache.atlas.type.Constants.PENDING_TASKS_PROPERTY_KEY;
@@ -103,9 +104,10 @@ public abstract class DeleteHandlerV1 {
     private   final TaskManagement       taskManagement;
     private   final AtlasGraph           graph;
     private   final TaskUtil             taskUtil;
+    private final TagDAO tagDAO;
 
     public DeleteHandlerV1(AtlasGraph graph, AtlasTypeRegistry typeRegistry, boolean shouldUpdateInverseReference, boolean softDelete,
-                           TaskManagement taskManagement, EntityGraphRetriever entityRetriever) {
+                           TaskManagement taskManagement, EntityGraphRetriever entityRetriever, TagDAO tagDAO) {
         this.typeRegistry                  = typeRegistry;
         this.graphHelper                   = new GraphHelper(graph);
         this.entityRetriever               = entityRetriever;
@@ -114,6 +116,7 @@ public abstract class DeleteHandlerV1 {
         this.taskManagement                = taskManagement;
         this.graph                         = graph;
         this.taskUtil                      = new TaskUtil(graph);
+        this.tagDAO = tagDAO;
     }
 
     /**
@@ -155,8 +158,17 @@ public abstract class DeleteHandlerV1 {
         // Delete traits and vertices.
         for (AtlasVertex deletionCandidateVertex : deletionCandidateVertices) {
             RequestContext.get().getDeletedEdgesIds().clear();
-            // TODO : create a V2 for this deleteClassifications which does not create these many redundant tasks
-            deleteAllClassifications(deletionCandidateVertex);
+            // TODO for Hritik Verify below
+            if(getJanusOptimisationEnabled()) {
+                List<AtlasClassification> classifications = entityRetriever.getAllClassifications_V2(deletionCandidateVertex);
+                classifications.forEach(classification -> {
+                    createAndQueueTaskWithoutCheckV2(CLASSIFICATION_PROPAGATION_DELETE, deletionCandidateVertex, null, classification.getTypeName());
+                });
+
+            } else {
+                deleteAllClassifications(deletionCandidateVertex);
+            }
+
             deleteTypeVertex(deletionCandidateVertex, isInternalType(deletionCandidateVertex));
 
             if (DEFERRED_ACTION_ENABLED) {
@@ -442,23 +454,30 @@ public abstract class DeleteHandlerV1 {
     }
 
     private void addTagPropagation(AtlasVertex fromVertex, AtlasVertex toVertex, AtlasEdge edge) throws AtlasBaseException {
-        final List<AtlasVertex> classificationVertices = getPropagationEnabledClassificationVertices(fromVertex);
-        String                  relationshipGuid       = getRelationshipGuid(edge);
-
-        if (taskManagement != null && DEFERRED_ACTION_ENABLED) {
-            for (AtlasVertex classificationVertex : classificationVertices) {
-                createAndQueueTask(CLASSIFICATION_PROPAGATION_ADD, toVertex, classificationVertex.getIdForDisplay(), getTypeName(classificationVertex), relationshipGuid);
-            }
+        //below needs to be forked ?
+        if(getJanusOptimisationEnabled()) {
+            // foreground
+            // classifcationTypeName can be empty
+            createAndQueueTaskWithoutCheckV2(CLASSIFICATION_PROPAGATION_ADD, fromVertex, toVertex, "");
         } else {
-            final List<AtlasVertex> propagatedEntityVertices = CollectionUtils.isNotEmpty(classificationVertices) ? entityRetriever.getIncludedImpactedVerticesV2(toVertex, relationshipGuid) : null;
+            final List<AtlasVertex> classificationVertices = getPropagationEnabledClassificationVertices(fromVertex);
+            String relationshipGuid = getRelationshipGuid(edge);
 
-            if (CollectionUtils.isNotEmpty(propagatedEntityVertices)) {
-                if (LOG.isDebugEnabled()) {
-                    LOG.debug("Propagate {} tags: from {} entity to {} entities", classificationVertices.size(), getTypeName(fromVertex), propagatedEntityVertices.size());
-                }
-
+            if (taskManagement != null && DEFERRED_ACTION_ENABLED) {
                 for (AtlasVertex classificationVertex : classificationVertices) {
-                    addTagPropagation(classificationVertex, propagatedEntityVertices);
+                    createAndQueueTask(CLASSIFICATION_PROPAGATION_ADD, toVertex, classificationVertex.getIdForDisplay(), getTypeName(classificationVertex), relationshipGuid);
+                }
+            } else {
+                final List<AtlasVertex> propagatedEntityVertices = CollectionUtils.isNotEmpty(classificationVertices) ? entityRetriever.getIncludedImpactedVerticesV2(toVertex, relationshipGuid) : null;
+
+                if (CollectionUtils.isNotEmpty(propagatedEntityVertices)) {
+                    if (LOG.isDebugEnabled()) {
+                        LOG.debug("Propagate {} tags: from {} entity to {} entities", classificationVertices.size(), getTypeName(fromVertex), propagatedEntityVertices.size());
+                    }
+
+                    for (AtlasVertex classificationVertex : classificationVertices) {
+                        addTagPropagation(classificationVertex, propagatedEntityVertices);
+                    }
                 }
             }
         }
@@ -1352,17 +1371,19 @@ public abstract class DeleteHandlerV1 {
         RequestContext.get().queueTask(task);
     }
 
-    public void createAndQueueTaskWithoutCheckV2(String taskType, AtlasVertex entityVertex, String classificationTypeName) {
+    public void createAndQueueTaskWithoutCheckV2(String taskType, AtlasVertex fromVertex, AtlasVertex toVertex, String classificationTypeName) {
         String              currentUser = RequestContext.getCurrentUser();
-        String              entityGuid  = GraphHelper.getGuid(entityVertex);
+        String              entityGuid  = GraphHelper.getGuid(fromVertex);
+        String              toEntityGuid  = toVertex != null ? GraphHelper.getGuid(toVertex) : null;
 
         Map<String, Object> taskParams  = new HashMap<>() {{
             put(PARAM_ENTITY_GUID, entityGuid);
             put(TASK_CLASSIFICATION_TYPENAME, classificationTypeName);
+            put(PARAM_TO_ENTITY_GUID, toEntityGuid);
         }};
 
         AtlasTask task = taskManagement.createTaskV2(taskType, currentUser, taskParams, classificationTypeName, entityGuid);
-        AtlasGraphUtilsV2.addEncodedProperty(entityVertex, PENDING_TASKS_PROPERTY_KEY, task.getGuid());
+        AtlasGraphUtilsV2.addEncodedProperty(fromVertex, PENDING_TASKS_PROPERTY_KEY, task.getGuid());
         RequestContext.get().queueTask(task);
     }
